+++ WorldOfCubeUE/src/org/worldOfCube/client/util/ShortPosition.java
@@ -1,21 +1,42 @@
+/*
+ * Copyright (c) 2012 matheusdev
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to
+ * deal in the Software without restriction, including without limitation the
+ * rights to use, copy, modify, merge, publish, distribute, sublicense, and/or
+ * sell copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
 package org.worldOfCube.client.util;
 
 public class ShortPosition {
-	
+
 	public static final short X = (short) 0xFF00;
 	public static final short Y = 0x00FF;
 	public static final int X_OFF = 8;
-	
+
 	public static short set(byte x, byte y) {
-		return (short) (((short)x << X_OFF) | ((short)y));
+		return (short) ((x << X_OFF) | (y));
 	}
-	
+
 	public static byte getX(short pos) {
 		return (byte)((pos & X) >> X_OFF);
 	}
-	
+
 	public static byte getY(short pos) {
 		return (byte)(pos & Y);
 	}
-	
+
 }
