+++ WorldOfCubeUE/src/org/worldOfCube/client/util/bresenhamline/Placer.java
@@ -1,3 +1,24 @@
+/*
+ * Copyright (c) 2012 matheusdev
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to
+ * deal in the Software without restriction, including without limitation the
+ * rights to use, copy, modify, merge, publish, distribute, sublicense, and/or
+ * sell copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
 package org.worldOfCube.client.util.bresenhamline;
 
 import org.worldOfCube.client.logic.chunks.World;
@@ -51,7 +72,7 @@ public final class Placer {
 			}
 		}
 	}
-	
+
 	public static void circle(int rx, int ry, double dist, PlaceAction pa, World world) {
 		int beginx = rx-(int)dist;
 		int beginy = ry-(int)dist;
@@ -61,11 +82,11 @@ public final class Placer {
 		beginy = Math.max(0, beginy);
 		endx = Math.min(world.totalBlocks, endx);
 		endy = Math.min(world.totalBlocks, endy);
-		double drx = (double)rx;
-		double dry = (double)ry;
+		double drx = rx;
+		double dry = ry;
 		for (int x = beginx; x < endx; x++) {
 			for (int y = beginy; y < endy; y++) {
-				if (Distance.get(drx, dry, (double)x, (double)y) <= dist) {
+				if (Distance.get(drx, dry, x, y) <= dist) {
 					pa.action(x, y);
 				}
 			}
