+++ WorldOfCubeUE/src/org/worldOfCube/client/util/vecmath/FastMath.java
@@ -1,3 +1,24 @@
+/*
+ * Copyright (c) 2012 matheusdev
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to
+ * deal in the Software without restriction, including without limitation the
+ * rights to use, copy, modify, merge, publish, distribute, sublicense, and/or
+ * sell copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
 package org.worldOfCube.client.util.vecmath;
 
 public class FastMath {
@@ -8,7 +29,7 @@ public class FastMath {
 	private static final int ATAN2_DIM = (int) Math.sqrt(ATAN2_COUNT);
 
 	private static final double INV_ATAN2_DIM_MINUS_1 = 1.0f / (ATAN2_DIM - 1);
-	private static final double DEG = 180.0f / (double) Math.PI;
+	private static final double DEG = 180.0f / Math.PI;
 
 	private static final double[] atan2 = new double[ATAN2_COUNT];
 
@@ -18,7 +39,7 @@ public class FastMath {
 				double x0 = (double) i / ATAN2_DIM;
 				double y0 = (double) j / ATAN2_DIM;
 
-				atan2[j * ATAN2_DIM + i] = (double) Math.atan2(y0, x0);
+				atan2[j * ATAN2_DIM + i] = Math.atan2(y0, x0);
 			}
 		}
 	}
@@ -32,7 +53,7 @@ public class FastMath {
 	}
 
 	public static final double atan2DegStrict(double y, double x) {
-		return (double) Math.atan2(y, x) * DEG;
+		return Math.atan2(y, x) * DEG;
 	}
 
 	public static final double atan2(double y, double x) {
